0314変更分、マイグレーション必須！

+++ src/apps/contracts/models.py
@@ -72,7 +72,7 @@ class FileupDetail(models.Model):
     # ワンタイムパスワード設定
     one_time_pw = models.BooleanField('ワンタイムパスワード設定可否', null=True, blank=True)
     # 一回の共有に対する最大ファイル容量(MBで計算　1024=1GB)
-    maxsize_every_share = models.IntegerField('1共有毎の最大ファイル容量', null=True, blank=True)
+    maxsize_every_share = models.BigIntegerField('1共有毎の最大ファイル容量', null=True, blank=True)
 """
 契約テーブル
 """
